+++ examples/data-sources/dedalus_machine/data-source.tf
@@ -1,3 +1,3 @@
 data "dedalus_machine" "example_machine" {
-  machine_id = "dm-3"
+  machine_id = "dm-ecc2efdd-ddfa-31a9-c6f1-b833d337aa7c"
 }