+++ examples/resources/dedalus_machine_preview/import.sh
@@ -1,1 +1,0 @@
-$ terraform import dedalus_machine_preview.example '<machine_id>/<preview_id>'